+++ package.json
@@ -4,9 +4,10 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "npx playwright test --project=chromium",
+    "test:chromium": "npx playwright test --project=chromium",
     "test:firefox": "npx playwright test --project=firefox",
-    "test:webkit": "npx playwright test --project=webkit"
+    "test:webkit": "npx playwright test --project=webkit",
+    "play-report": "npx playwright show-report"
   },
   "keywords": [],
   "author": "",
